RL-Bitcoin-trading-bot_7/requirements.txt
numpy
tensorflow==2.3.1
tensorflow-gpu==2.3.1
opencv-python
matplotlib
tensorboardx
pandas
mplfinance
ta
bitfinex-tencars
seaborn
